Add m3_LinkRawTableFunction to extensions
A common way to pass a callback into as wasm program is to insert
the callback into the indirect function call table. (See
library_addfunction.js from emscripten).

In javascript, the type of the function doesnt need to be specified,
call_indirect will always just work on the index if its an external
function (assuming the table is type anyfunc). Wasm3 calls raw
functiosn via a wasm trampoline so here we need to pass a function
sig for the trampoline.